Extract GameDock from Game.tsx.
